Add DI DNS implementation
