+++ src/setup.ts
@@ -2,10 +2,12 @@ import { Endpoints, setup as internalSetup, UserMessages } from "./setup/interna
 
 import { Implementation as AuthImplementation } from "./auth/implementation/types.js"
 import { Implementation as CryptoImplementation } from "./crypto/implementation/types.js"
+import { Implementation as DnsImplementation } from "./dns/implementation/types.js"
 import { Implementation as StorageImplementation } from "./storage/implementation/types.js"
 
 import * as authImpl from "./auth/implementation.js"
 import * as cryptoImpl from "./crypto/implementation.js"
+import * as dnsImpl from "./dns/implementation.js"
 import * as storageImpl from "./storage/implementation.js"
 
 
@@ -56,10 +58,12 @@ export function endpoints(e: Partial<Endpoints>): Endpoints {
 export function setImplementations(opts: {
   auth?: Partial<AuthImplementation>
   crypto?: Partial<CryptoImplementation>
+  dns?: Partial<DnsImplementation>
   storage?: Partial<StorageImplementation>
 }) {
   if (opts.auth) authImpl.set(opts.auth)
   if (opts.crypto) cryptoImpl.set(opts.crypto)
+  if (opts.dns) dnsImpl.set(opts.dns)
   if (opts.storage) storageImpl.set(opts.storage)
 }
